Add User DTO and Create User Handler
This commit adds a new User DTO and implements a Create User handler in the web adapter. The User DTO is used to bind incoming JSON data to the application User struct. The Create User handler decodes the JSON data, creates a new user using the UserService, and encodes the user as JSON in the response.